requirements.txt
google-cloud-pubsub==2.4.1
matplotlib==3.4.1
memory-profiler==0.58.0
